+++ src/tests/index.c
@@ -5,7 +5,7 @@
 
 int main() {
     SString *str1 = CStringToSSTring("dromedario");
-    size_t index1 = indexof(str1, "eda");
+    int index1 = indexof(str1, "eda");
     if (index1 != 4) {
         printf("Not Ok, index should 4, but got %d\n", index1);
         return 0;
@@ -13,7 +13,7 @@ int main() {
     printf("index Ok, 4.\n");
 
     SString *str2 = CStringToSSTring("fooafoodfoofoo");
-    size_t index2 = indexof(str2, "foo");
+    int index2 = indexof(str2, "foo");
     if (index2 != 0) {
         printf("Not Ok, index should 0, but got %d\n", index2);
         return 0;
@@ -21,7 +21,7 @@ int main() {
     printf("index Ok, 0.\n");
 
     SString *str3 = CStringToSSTring("asdfbar");
-    size_t index3 = indexof(str3, "bar");
+    int index3 = indexof(str3, "bar");
     if (index3 != 4) {
         printf("Not Ok, index should 4, but got %d\n", index3);
         return 0;
@@ -29,7 +29,7 @@ int main() {
     printf("index Ok, 5.\n");
 
     SString *str4 = CStringToSSTring("bar");
-    size_t index4 = indexof(str4, "r");
+    int index4 = indexof(str4, "r");
     if (index4 != 2) {
         printf("Not Ok, index should 2, but got %d\n", index4);
         return 0;
@@ -37,7 +37,7 @@ int main() {
     printf("index Ok, 2.\n");
 
     SString *str5 = CStringToSSTring("aaaaabbbaa");
-    size_t index5 = indexof(str5, "c");
+    int index5 = indexof(str5, "c");
     if (index5 != -1) {
         printf("Not Ok, index should -1, but got %d\n", index5);
         return 0;
